JolinPluto

Documentation for JolinPluto.

JolinPluto.embedLargeHTML

JolinPluto.is_running_in_pluto_process

JolinPluto.@Channel

JolinPluto.@authorize_aws

JolinPluto.@clipboard_image_to_clipboard_html

JolinPluto.@get_jwt

JolinPluto.@give_me_register_cleanup_function

JolinPluto.@give_me_rerun_cell_function

JolinPluto.@output_below

JolinPluto.@repeat_at

JolinPluto.@repeat_run

JolinPluto.@repeat_take!

JolinPluto.embedLargeHTML — Method

embedLargeHTML(html_string; width, height)

For large HTML code a simple iframe is not enough. This function will help.

Example

JolinPluto.embedLargeHTML(read("figure.html", String); width="100%", height=400)

JolinPluto.is_running_in_pluto_process — Method

is_running_in_pluto_process()

This doesn't mean we're in a Pluto cell, e.g. can use @bind and hooks goodies. It only means PlutoRunner is available (and at a version that technically supports hooks)

JolinPluto.@Channel — Macro

channel = @Channel(10) do ch
    for i in 1:10
        put!(ch, i)
        sleep(1)
    end
end

Like normal Channel, with the underlying task being interrupted as soon as the Pluto cell is deleted.

JolinPluto.@authorize_aws — Macro

@authorize_aws(role_arn; audience="")

Assume role via web identity. How to define such a role can be found here https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-role.html#cli-configure-role-oidc

CAUTION: Please note that the semicolon is really important! @authorize_aws(role_arn, audience="myaudience") won't work as of now.

JolinPluto.@clipboard_image_to_clipboard_html — Macro

@clipboardimagetoclipboardhtml

Creates a little textfield where you can paste images. These images are then transformed to self-containing html img tags and copied back to the clipboard to be entered somewhere in Pluto.

JolinPluto.@get_jwt — Macro

@get_jwt
@get_jwt "exampleaudience"

Creates a JSON Web Token which can be used for authentication at common cloud providers.

On cloud.jolin.io the token will be issued and signed by cloud.jolin.io, on Github Actions (used for automated tests), a respective github token is returned.

JolinPluto.@give_me_register_cleanup_function — Macro

@give_me_register_cleanup_function()

⚠️ Don't use this directly!! if you think you need it, you need @use_effect.

Used inside a Pluto cell this will resolve to a function that call be called with yet another function, and then will call that function when the cell gets explicitly re-run. ("Explicitly re-run" meaning all @use_refs get cleared, for example). Outside a Pluto cell it will throw an error.

JolinPluto.@give_me_rerun_cell_function — Macro

@give_me_rerun_cell_function()

⚠️ Don't use this directly!! if you think you need it, you need @use_state.

Used inside a Pluto cell this will resolve to a function that, when called, will cause the cell to be re-run (in turn re-running all dependent cells). Outside a Pluto cell it will throw an error.

JolinPluto.@output_below — Macro

@output_below

Reverse input output, first input then output. When removing the cell with @output_below, the order is reversed again.

JolinPluto.@repeat_at — Macro

@repeat_at(ceil(now(), Second(10)), init=:wait) do t
	# code to be returned repeatedly
	rand(), t
end

When run inside Pluto it will rerun the function on the next specified time, again and again.

Keyword Arguments

init specifies what to do at first run or re-evaluation caused by standard reactivity. You can specify any code or function call (e.g. init=nothing, or init=myinit()). In addition there are two special values :wait und :run. :wait (default) will wait for the next time and then run the code. :run will run the code immediately without waiting.

JolinPluto.@repeat_run — Macro

@repeat_run(expr_used_for_init_and_repeated_execution)
@repeat_run(expr_for_init, expr_for_repetition)

Repeat some long expression. It will run offline in a separate Task so that interactivity is preserved.

Optionally, specify another expression for initialization. Initialization will also be triggered if the cell is re-evaluated because some dependent cell or bond changed.

JolinPluto.@repeat_take! — Macro

nextvalue = @repeat_take! channel

This will repeatedly fetch for the next element from the given channel, re-evaluating the cell each time a new value arrives.