JolinPluto

Documentation for JolinPluto.

JolinPluto.Setter

JolinPluto.ChannelPluto

JolinPluto._split_macro_args

JolinPluto.authorize_aws

JolinPluto.cell_ids_create_wrapper

JolinPluto.cell_ids_push!

JolinPluto.embedLargeHTML

JolinPluto.get_jwt

JolinPluto.is_running_in_pluto_process

JolinPluto.output_below

JolinPluto.repeat_at

JolinPluto.repeat_put_at

JolinPluto.repeat_queueget

JolinPluto.repeat_run

JolinPluto.repeat_take!

JolinPluto.@Channel

JolinPluto.@authorize_aws

JolinPluto.@cell_ids_create_wrapper

JolinPluto.@cell_ids_push!

JolinPluto.@clipboard_image_to_clipboard_html

JolinPluto.@get_jwt

JolinPluto.@give_me_register_cleanup_function

JolinPluto.@give_me_rerun_cell_function

JolinPluto.@give_me_the_pluto_cell_id

JolinPluto.@output_below

JolinPluto.@repeat_at

JolinPluto.@repeat_run

JolinPluto.@repeat_take!

JolinPluto.Setter — Type

Setter()
Setter("initial_value")

Creates a pluto interactivity which separates the setter cell from the state cell.

Usage

set_a = Setter("initial_value")

# in another cell, extract the inner state from the setter
# such that updates will rerun this cell
a = @get set_a

# in yet another cell use `set_a`
set_a("new_value")

# or use a function syntax to easily access the previous value
set_a() do prev_a
    "$prev_a!!"
end

JolinPluto.ChannelPluto — Method

channel = ChannelPluto(10) do ch
    for i in 1:10
        put!(ch, i)
        sleep(1)
    end
end

Like normal Channel, with the underlying task being interrupted as soon as the Pluto cell is deleted.

JolinPluto._split_macro_args — Method

separates args and kwargs (kwargs are returned as Dict)

JolinPluto.authorize_aws — Method

authorize_aws(role_arn; audience="")

Assume role via web identity. How to define such a role can be found here https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-role.html#cli-configure-role-oidc

CAUTION: Please note that the semicolon is really important! @authorize_aws(role_arn, audience="myaudience") won't work as of now.

JolinPluto.cell_ids_create_wrapper — Method

cell_ids_wrapper = cell_ids_create_wrapper()

Creates a wrapper around a Set of cell_ids so that they can be added (and removed) seamlessly and automatically from other cells.

Usage

cell_ids_wrapper = cell_ids_create_wrapper()

# in another cell access the cellids and e.g. print the url suffix
cell_ids = get(cell_ids_wrapper)
print(join("&isolated_cell_id=$id" for id in cell_ids))

# in yet another cell add its cell_id
cell_ids_push!(cell_ids_wrapper)

JolinPluto.cell_ids_push! — Method

cell_ids_push!(cell_ids_wrapper)

Adds the cell's cell-id to the given cellidswrapper. This automatically handles retriggering of cells as needed.

Also cleanup is handled, i.e. that the cell-id is removed again if this cell is deleted.

JolinPluto.embedLargeHTML — Method

embedLargeHTML(html_string; width, height)

For large HTML code a simple iframe is not enough. This function will help.

Example

JolinPluto.embedLargeHTML(read("figure.html", String); width="100%", height=400)

JolinPluto.get_jwt — Function

get_jwt()
get_jwt("exampleaudience")

Creates a JSON Web Token which can be used for authentication at common cloud providers.

On cloud.jolin.io the token will be issued and signed by cloud.jolin.io, on Github Actions (used for automated tests), a respective github token is returned.

JolinPluto.is_running_in_pluto_process — Method

is_running_in_pluto_process()

This doesn't mean we're in a Pluto cell, e.g. can use @bind and hooks goodies. It only means PlutoRunner is available (and at a version that technically supports hooks)

JolinPluto.output_below — Method

output_below()

Reverse input output, first input then output. When removing the cell with @output_below, the order is reversed again.

JolinPluto.repeat_at — Method

repeat_at(t -> (rand(), t), ceil(now(), Second(10)), init=:wait)

repeat_at(ceil(now(), Second(10)), init=:wait) do t
	# code to be returned repeatedly
	rand(), t
end

When run inside Pluto it will rerun the function on the next specified time, again and again.

Keyword Arguments

init specifies what to do at first run or re-evaluation caused by standard reactivity. You can specify any function (e.g. init=myinit_func)or one of the two special values :wait und :run. :wait (default) will wait for the next time and then run the code. :run will run the code immediately without waiting.

JolinPluto.repeat_put_at — Method

repeat_put_at(channel, (time) -> value, nexttime)

Use this function from other languages which cannot safely run put!(channel, value). This ensures that the put! is called from julia, circumventing segmentation faults because of async switch inside another-language-function.

JolinPluto.repeat_queueget — Function

repeat_queueget(python_queue_threaded)

Will repeatedly get elements from the queue and trigger a rerun of the current cell. Outside pluto it will just wait for the first element to arrive and return that.

IMPORTANT: This function is only available if import PythonCall was executed before import JolinPluto.

JolinPluto.repeat_run — Function

repeat_run(function_used_for_init_and_repeated_execution)
repeat_run(function_for_init, function_for_repetition)

Repeat some long expression. It will run offline in a separate Task so that interactivity is preserved.

Optionally, specify another expression for initialization. Initialization will also be triggered if the cell is re-evaluated because some dependent cell or bond changed.

JolinPluto.repeat_take! — Method

nextvalue = repeat_take!(channel)

This will repeatedly fetch for the next element from the given channel, re-evaluating the cell each time a new value arrives.

JolinPluto.@Channel — Macro

channel = @Channel(10) do ch
    for i in 1:10
        put!(ch, i)
        sleep(1)
    end
end

Like normal Channel, with the underlying task being interrupted as soon as the Pluto cell is deleted.

JolinPluto.@authorize_aws — Macro

@authorize_aws(role_arn; audience="")

Assume role via web identity. How to define such a role can be found here https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-role.html#cli-configure-role-oidc

CAUTION: Please note that the semicolon is really important! @authorize_aws(role_arn, audience="myaudience") won't work as of now.

JolinPluto.@cell_ids_create_wrapper — Macro

cell_ids_wrapper = @cell_ids_create_wrapper()

Creates a wrapper around a Set of cell_ids so that they can be added (and removed) seamlessly and automatically from other cells.

Usage

cell_ids_wrapper = @cell_ids_create_wrapper()

# in another cell access the cellids and e.g. print the url suffix
cell_ids = @get cell_ids_wrapper
print(join("&isolated_cell_id=$id" for id in cell_ids))

# in yet another cell add its cell_id
@cell_ids_push! cell_ids_wrapper

JolinPluto.@cell_ids_push! — Macro

@cell_ids_push! cell_ids_wrapper

Adds the cell's cell-id to the given cellidswrapper. This automatically handles retriggering of cells as needed.

Also cleanup is handled, i.e. that the cell-id is removed again if this cell is deleted.

JolinPluto.@clipboard_image_to_clipboard_html — Macro

@clipboardimagetoclipboardhtml

Creates a little textfield where you can paste images. These images are then transformed to self-containing html img tags and copied back to the clipboard to be entered somewhere in Pluto.

JolinPluto.@get_jwt — Macro

@get_jwt
@get_jwt "exampleaudience"

Creates a JSON Web Token which can be used for authentication at common cloud providers.

On cloud.jolin.io the token will be issued and signed by cloud.jolin.io, on Github Actions (used for automated tests), a respective github token is returned.

JolinPluto.@give_me_register_cleanup_function — Macro

@give_me_register_cleanup_function()

⚠️ Don't use this directly!! if you think you need it, you need @use_effect.

Used inside a Pluto cell this will resolve to a function that call be called with yet another function, and then will call that function when the cell gets explicitly re-run. ("Explicitly re-run" meaning all @use_refs get cleared, for example). Outside a Pluto cell it will throw an error.

JolinPluto.@give_me_rerun_cell_function — Macro

@give_me_rerun_cell_function()

⚠️ Don't use this directly!! if you think you need it, you need @use_state.

Used inside a Pluto cell this will resolve to a function that, when called, will cause the cell to be re-run (in turn re-running all dependent cells). Outside a Pluto cell it will throw an error.

JolinPluto.@give_me_the_pluto_cell_id — Macro

@give_me_the_pluto_cell_id()

⚠️ Don't use this directly!! if you think you need it, you might actually need @use_did_deps_change([]) but even that is unlikely.

Used inside a Pluto cell this will resolve to the current cell UUID. Outside a Pluto cell it will throw an error.

JolinPluto.@output_below — Macro

@output_below

Reverse input output, first input then output. When removing the cell with @output_below, the order is reversed again.

JolinPluto.@repeat_at — Macro

@repeat_at(ceil(now(), Second(10)), init=:wait) do t
	# code to be returned repeatedly
	rand(), t
end

@repeat_at(ceil(now(), Second(10)), init=:wait)

When run inside Pluto it will rerun the function on the next specified time, again and again.

Keyword Arguments

init specifies what to do at first run or re-evaluation caused by standard reactivity. You can specify any code or function call (e.g. init=nothing, or init=myinit()). In addition there are two special values :wait und :run. :wait (default) will wait for the next time and then run the code. :run will run the code immediately without waiting.

JolinPluto.@repeat_run — Macro

@repeat_run(expr_used_for_init_and_repeated_execution)
@repeat_run(expr_for_init, expr_for_repetition)

Repeat some long expression. It will run offline in a separate Task so that interactivity is preserved.

Optionally, specify another expression for initialization. Initialization will also be triggered if the cell is re-evaluated because some dependent cell or bond changed.

JolinPluto.@repeat_take! — Macro

nextvalue = @repeat_take! channel

This will repeatedly fetch for the next element from the given channel, re-evaluating the cell each time a new value arrives.